feat(calculate) : 덧셈 기능 및 출력 기능 구현
- 덧셈 범위 long 초과 시 IllegalArgumentException throw

+++ src/main/java/calculator/CalculatorImpl.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 public interface CalculatorImpl {
     void run();
-    int calculate(List<Long> numbers);
+    long calculate(List<Long> numbers);
     void print(long result);
 }
